refactor(eve): group HITL input resolvers
